Package name refactoring
