+++ src/test/java/com/toolable/notes/stub/BaseStubTest.java
@@ -1,9 +1,9 @@
-package com.toolable.notesstub.stub;
+package com.toolable.notes.stub;
 
 import com.google.common.base.Preconditions;
-import com.toolable.notesstub.exceptions.NotImplementedException;
-import com.toolable.notesstub.exceptions.RecycledObjectException;
-import com.toolable.notesstub.exceptions.UnexpectedException;
+import com.toolable.notes.stub.exceptions.NotImplementedException;
+import com.toolable.notes.stub.exceptions.RecycledObjectException;
+import com.toolable.notes.stub.exceptions.UnexpectedException;
 import lotus.domino.Base;
 import lotus.domino.NotesException;
 import org.junit.Assert;
